Split config into a separate file
