+++ dispatcher.go
@@ -37,112 +37,13 @@ import (
 	"go.uber.org/yarpc/internal/request"
 	intsync "go.uber.org/yarpc/internal/sync"
 
-	"github.com/opentracing/opentracing-go"
-	"github.com/prometheus/client_golang/prometheus"
-	"github.com/uber-go/tally"
 	"go.uber.org/multierr"
 	"go.uber.org/zap"
-	"go.uber.org/zap/zapcore"
 )
 
 // Default sleep between pushes to Tally metrics. At some point, we may want
 // this to be configurable.
 const _tallyPushInterval = 500 * time.Millisecond
-
-// LoggingConfig describes how logging should be configured.
-type LoggingConfig struct {
-	// Supplies a logger for the dispatcher. By default, no logs are
-	// emitted.
-	Zap *zap.Logger
-	// If supplied, the ContextExtractor is used to log request-scoped
-	// information carried on the context (e.g., trace and span IDs).
-	ContextExtractor func(context.Context) zapcore.Field
-}
-
-func (c LoggingConfig) logger(name string) *zap.Logger {
-	if c.Zap == nil {
-		return zap.NewNop()
-	}
-	return c.Zap.Named("yarpc").With(
-		// Use a namespace to prevent key collisions with other libraries.
-		zap.Namespace("yarpc"),
-		zap.String("dispatcher", name),
-	)
-}
-
-func (c LoggingConfig) extractor() observerware.ContextExtractor {
-	if c.ContextExtractor == nil {
-		return observerware.NewNopContextExtractor()
-	}
-	return observerware.ContextExtractor(c.ContextExtractor)
-}
-
-// MetricsConfig describes how telemetry should be configured.
-type MetricsConfig struct {
-	// Tally scope used for pushing to M3 or StatsD-based systems. By
-	// default, metrics are collected in memory but not pushed.
-	Tally tally.Scope
-}
-
-func (c MetricsConfig) registry(name string, logger *zap.Logger) (*pally.Registry, context.CancelFunc) {
-	r := pally.NewRegistry(
-		pally.Labeled(pally.Labels{
-			"component":  "yarpc",
-			"dispatcher": pally.ScrubLabelValue(name),
-		}),
-		// Also expose all YARPC metrics via the default Prometheus registry.
-		pally.Federated(prometheus.DefaultRegisterer),
-	)
-
-	if c.Tally == nil {
-		return r, nil
-	}
-
-	stop, err := r.Push(c.Tally, _tallyPushInterval)
-	if err != nil {
-		logger.Error("Failed to start pushing metrics to Tally.", zap.Error(err))
-		return r, nil
-	}
-	return r, stop
-}
-
-// Config specifies the parameters of a new Dispatcher constructed via
-// NewDispatcher.
-type Config struct {
-	// Name of the service. This is the name used by other services when
-	// making requests to this service.
-	Name string
-
-	// Inbounds define how this service receives incoming requests from other
-	// services.
-	//
-	// This may be nil if this service does not receive any requests.
-	Inbounds Inbounds
-
-	// Outbounds defines how this service makes requests to other services.
-	//
-	// This may be nil if this service does not send any requests.
-	Outbounds Outbounds
-
-	// Inbound and Outbound Middleware that will be applied to all incoming
-	// and outgoing requests respectively.
-	//
-	// These may be nil if there is no middleware to apply.
-	InboundMiddleware  InboundMiddleware
-	OutboundMiddleware OutboundMiddleware
-
-	// Tracer is deprecated. The dispatcher does nothing with this propery.
-	Tracer opentracing.Tracer
-
-	// RouterMiddleware is middleware to control how requests are routed.
-	RouterMiddleware middleware.Router
-
-	// Logging configures logging.
-	Logging LoggingConfig
-
-	// Metrics configures telemetry.
-	Metrics MetricsConfig
-}
 
 // Inbounds contains a list of inbound transports. Each inbound transport
 // specifies a source through which incoming requests are received.
